added variables
