+++ package.json
@@ -45,7 +45,6 @@
         "@typescript-eslint/parser": "5.36.1",
         "antd": "4.21.6",
         "babel-loader": "8.2.5",
-        "babel-polyfill": "6.26.0",
         "babel-register": "6.26.0",
         "canvas": "2.10.0",
         "cross-env": "7.0.3",
@@ -83,10 +82,8 @@
         "typescript": "4.7.4",
         "url": "0.11.0",
         "url-loader": "4.1.1",
-        "webpack": "5.73.0",
         "webpack-cli": "4.10.0",
         "webpack-dev-server": "4.9.3",
-        "webpack-hot-middleware": "2.25.2",
         "webpack-node-externals": "3.0.0",
         "workbox-precaching": "6.5.4",
         "workbox-webpack-plugin": "6.5.4",
@@ -95,6 +92,7 @@
     "dependencies": {
         "@reduxjs/toolkit": "1.8.3",
         "antd": "4.21.6",
+        "babel-polyfill": "6.26.0",
         "express": "4.18.1",
         "htmlescape": "1.1.1",
         "merge-options": "3.0.4",
@@ -107,6 +105,9 @@
         "react-redux": "7.2.8",
         "react-router": "6.3.0",
         "react-router-dom": "6.3.0",
+        "webpack": "5.73.0",
+        "webpack-dev-middleware": "5.3.3",
+        "webpack-hot-middleware": "2.25.2",
         "workbox-window": "6.5.4"
     },
     "engines": {
