+++ src/components/form/UserinfoForm/UserinfoForm.scss
@@ -16,7 +16,7 @@
             @include font-style(
                 $font-size: 16px,
                 $letter-spacing: normal,
-                $color: $white,
+                $color: rgb($white),
                 $font-family: Ubuntu,
                 $font-weight: 300
             );
@@ -27,7 +27,7 @@
             height: 46px;
 
             span {
-                @include font-style($font-size: 33px, $color: $white, $font-family: VT323);
+                @include font-style($font-size: 33px, $color: rgb($white), $font-family: VT323);
             }
         }
     }
