+++ src/components/icons/ArrowSend/ArrowSend.scss
@@ -3,5 +3,5 @@
 .arrow-send-icon {
     vertical-align: middle;
 
-    fill: $red;
+    fill: rgb($red);
 }
