+++ src/components/icons/CloseButton/CloseButton.scss
@@ -3,5 +3,5 @@
 .close-buton-icon {
     vertical-align: middle;
 
-    fill: $red;
+    fill: rgb($red);
 }
