+++ src/components/icons/Pencil/Pencil.scss
@@ -3,5 +3,5 @@
 .pencil-icon {
     vertical-align: middle;
 
-    fill: $red;
+    fill: rgb($red);
 }
