+++ src/components/icons/ReplyCounterArrow/ReplyCounterArrow.scss
@@ -3,5 +3,5 @@
 .reply-counter-icon {
     vertical-align: middle;
 
-    fill: $white;
+    fill: rgb($white);
 }
