+++ src/components/sections/HeaderBackButton/HeaderBackButton.scss
@@ -12,7 +12,7 @@
         padding: 0;
 
         font-size: $xxxl;
-        color: $red;
+        color: rgb($red);
         vertical-align: middle;
     }
 }
