+++ src/game/ui/PauseMenu/PauseMenu.scss
@@ -15,7 +15,7 @@
     width: 100vw;
     height: 100vh;
 
-    background-color: adjust-color($black-hilighted, $alpha: -0.5);
+    background-color: rgba($black-hilighted, 0.5);
 
     &.active {
         animation-name: scale-open;
@@ -31,7 +31,7 @@
         @include VT323;
 
         font-size: $xxxl;
-        color: $white;
+        color: rgb($white);
     }
 
     .ant-btn-primary {
