+++ src/pages/game/Game.scss
@@ -14,7 +14,7 @@
     width: 100vw;
     height: 100vh;
 
-    background-color: $green;
+    background-color: rgb($green);
 
     .game-page-map-container {
         position: relative;
