+++ src/pages/game/components/GameBackButton/GameBackButton.scss
@@ -18,26 +18,26 @@
         text-align: center;
         vertical-align: middle;
 
-        background-color: $black;
+        background-color: rgb($black);
         border: 0;
         border-radius: 20px;
         outline: 0;
-        box-shadow: 5px 5px 10px 0 adjust-color($black-hilighted, $alpha: -0.5);
+        box-shadow: 5px 5px 10px 0 rgba($black-hilighted, 0.5);
 
         transition: background-color 0.3s, transform 0.3s, box-shadow 0.3s;
 
         &:hover {
-            background-color: $black-hilighted;
+            background-color: rgb($black-hilighted);
         }
 
         &:active {
             transform: translate(5px, 5px);
 
-            box-shadow: 0 0 10px 0 adjust-color($black-hilighted, $alpha: -0.5);
+            box-shadow: 0 0 10px 0 rgba($black-hilighted, 0.5);
         }
 
         .arrow-back-icon {
-            fill: $white;
+            fill: rgb($white);
         }
     }
 }
