+++ src/pages/settings/Settings.scss
@@ -6,10 +6,10 @@
     margin-bottom: 40px;
     padding: 40px 120px;
 
-    background-color: $brown;
+    background-color: rgb($brown);
 
     * {
-        color: $white;
+        color: rgb($white);
     }
 
     .form-button,
@@ -17,7 +17,7 @@
         &:active,
         &:focus,
         &:hover {
-            box-shadow: inset 0 -5px 0 0 $black;
+            box-shadow: inset 0 -5px 0 0 rgb($black);
         }
     }
