+++ src/styles/antd/input.scss
@@ -5,25 +5,25 @@
 
     .ant-input,
     .ant-input-status-error {
-        @include font-style($font-size: 40px, $color: $white, $font-family: VT323);
+        @include font-style($font-size: 40px, $color: rgb($white), $font-family: VT323);
 
         cursor: pointer;
 
         padding: 24px 15px;
 
-        background: $black !important;
+        background: rgb($black) !important;
         border: 0;
         border-radius: 0 20px 0 0;
 
         transition: background-color 0.3s, box-shadow 0.3s;
 
         &:hover {
-            background: $black-hilighted;
+            background: rgb($black-hilighted);
         }
 
         &:active,
         &:focus {
-            box-shadow: inset 0 -5px 0 0 $brown, 0 0 0 100px $black-hilighted inset;
+            box-shadow: inset 0 -5px 0 0 rgb($brown), 0 0 0 100px rgb($black-hilighted) inset;
         }
 
         input { // это чтобы хром не перекрашивал все при автозаполнении
@@ -32,10 +32,10 @@
                 @include VT323(true);
 
                 /* stylelint-disable-next-line property-no-vendor-prefix */
-                -webkit-box-shadow: 0 0 0 100px $black inset;
-                box-shadow: 0 0 0 100px $black inset;
+                -webkit-box-shadow: 0 0 0 100px rgb($black) inset;
+                box-shadow: 0 0 0 100px rgb($black) inset;
 
-                -webkit-text-fill-color: $white;
+                -webkit-text-fill-color: rgb($white);
             }
         }
     }
@@ -47,7 +47,7 @@
         left: 11px;
 
         & label {
-            @include font-style($font-size: $m, $color: $white, $font-family: Ubuntu, $font-weight: 300);
+            @include font-style($font-size: $m, $color: rgb($white), $font-family: Ubuntu, $font-weight: 300);
 
             letter-spacing: normal;
 
@@ -76,7 +76,7 @@
         width: auto;
         min-height: 0;
 
-        color: $red;
+        color: rgb($red);
         white-space: nowrap;
     }
 }
