+++ src/styles/common.scss
@@ -14,5 +14,5 @@ body {
     font-weight: 400;
     font-style: normal;
     line-height: 13px;
-    color: $black;
+    color: rgb($black);
 }
