+++ src/styles/mixins.scss
@@ -43,7 +43,7 @@
 }
 @mixin font-style(
     $font-size: 20px,
-    $color: $black,
+    $color: rgb($black),
     $font-family: Ubuntu,
     $font-weight: 400,
     $line-height: 23px,
