+++ src/styles/variables.scss
@@ -1,10 +1,28 @@
+:root {
+    --black-color: 54, 39, 36;
+    --white-color: 250, 246, 236;
+    --red-color: 188, 48, 47;
+    --black-hilighted-color: 34, 34, 34;
+    --brown-color: 102, 61, 57;
+    --green-color: 117, 188, 47;
+}
+
+.alt-theme {
+    --black-color: 54, 39, 36;
+    --white-color: 135, 76, 98;
+    --red-color: 188, 48, 47;
+    --black-hilighted-color: 34, 34, 34;
+    --brown-color: 102, 61, 57;
+    --green-color: 117, 188, 47;
+}
+
 // COLORS
-$black: #362724;
-$white: #faf6ec;
-$red: #bc302f;
-$black-hilighted: #222;
-$brown: #663d39;
-$green: #75bc2f;
+$black: var(--black-color);
+$white: var(--white-color);
+$red: var(--red-color);
+$black-hilighted: var(--black-hilighted-color);
+$brown: var(--brown-color);
+$green: var(--green-color);
 
 //FONT-SIZES
 $normal: 1rem; // 13px
